Create separate modules for each KKT backend
