+++ test/KKT/Cholmod/cholmod.jl
@@ -6,12 +6,12 @@
     ])
 
     @testset "LDL" begin
-        kkt = KKT.setup(A, KKT.K2(), KKT.CholmodBackend())
+        kkt = KKT.setup(A, KKT.K2(), KKT.TlpCholmod.Backend())
         KKT.run_ls_tests(A, kkt)
     end
 
     @testset "Cholesky" begin
-        kkt = KKT.setup(A, KKT.K1(), KKT.CholmodBackend())
+        kkt = KKT.setup(A, KKT.K1(), KKT.TlpCholmod.Backend())
         KKT.run_ls_tests(A, kkt)
     end
 end
