+++ test/KKT/Dense/lapack.jl
@@ -5,7 +5,7 @@
                 1 0 1 0;
                 0 1 0 1
             ])
-            kkt = KKT.setup(A, KKT.K1(), KKT.DenseBackend())
+            kkt = KKT.setup(A, KKT.K1(), KKT.TlpDense.Backend())
             KKT.run_ls_tests(A, kkt)
         end
     end
