+++ test/KKT/LDLFactorizations/ldlfact.jl
@@ -5,7 +5,7 @@
                 1 0 1 0;
                 0 1 0 1
             ])
-            kkt = KKT.setup(A, KKT.K2(), KKT.LDLFactBackend())
+            kkt = KKT.setup(A, KKT.K2(), KKT.TlpLDLFact.Backend())
             KKT.run_ls_tests(A, kkt)
         end
     end
